VersionManager: Attach CVersionManagerDlg* through constructor
We could use AfxGetApp()->GetMainWnd() but this is overkill.
Just initialise it directly and we're fine.

The reason this doesn't work as-is is because AfxGetMainWnd()
fetches the window associated with the current thread.

Since packet requests are handled in the ReceiveWorkerThread which
is not associated with the window, it will fail.