VersionManager: Use temporary manual auth logic for now
Ideally we should have this in its own account database, or at the very least, rename the procedure.
ACCOUNT_LOGIN as-is, is the game implementation. We do not want to use this for login as it is incompatible.

As we're currently using a singular database (and we expect people to be using our database), and we have no
means of syncing this currently, we'll temporarily hack this to fetch and handle basic auth logic
without a procedure.