Restore N3UIWndBase.cpp to client project
It was only partially moved into the N3Base project for some reason, when this is a client wrapper.
I suspect the reason it was initially moved was because some of the controls (e.g. CN3UIArea) reference this,
but they continue to refer to the header in its original place so... I don't think the mission was accomplished there.

It doesn't really need to be building for tools and such though. It's intended to be game code.