Setup CodeQL
For this to work, it needs to build the entire thing. The split builds will confuse it if analysed with those.
Ideally this should use All.sln but this isn't setup yet, so it does it manually.

For now, it's triggered:
1. Manually via workflow_dispatch
2. Automatically once every week
3. Automatically in PRs via the run-analysis label, since there's not really a better way to enforce it in PRs.
   ready_for_review is close, but if it was never a draft in the first place, that won't trigger.
   It would need to be toggled.

I think this is fine. Most of the time it won't really need to run anyway, so it makes sense to just leave it.
The periodic automated scan should be sufficient.

Additionally, all 32-bit platform names are made consistent. Some were x86, some were Win32.
This keeps the builds consistent and avoids having to specify multiple platforms in codeql.yml.

The native configuration is actually Win32 (not sure where x86 came from), so I swapped it back
to Win32. I considered making it x86, but chose against it because Visual Studio tries to be clever
and inserts the x64 configuration, AND THEN because it's alphanumerically ordered, things default to x64.

So Win32 it is.