Default login server IP to 127.0.0.1 in debug builds
Since a default isn't otherwise provided with the assets,
I think defaulting this in debug builds is a fair compromise.

We should probably just create a default when it's pulled,
but for now this helps things to "just run".

There's no reason to keep this behaviour in release builds
since it's only intended for local development.

They can setup server.ini if they need to.