Simplify workflows
Go back to triggering build_all.yml for everything.
Only trigger the others if we explicitly change those workflows (or if we manually request it).

Additionally, in "all", rather than just running All.sln (and having no idea how many are actually left),
we run each of the individual solutions to break up the workflow's progress in a measurable way.