Scope timer to login scene + reset upon return to the scene
We scope the timer here to avoid redundant logic (although *very* minor) running
in every other scene.
The only reason it wasn't, was because we have both 1.098 and 1.298 scenes, and
they don't otherwise share any logic.

Since they only exist at compile time, we'll just move this back under the scene
specifically.

Additionally, Init() is called whenever the scene is entered, even when returning.
If we're returning, we should reset the timer for new connections, so this is the
simplest way to do that.

The point is to specifically remove the spam window between establishing a game
connection and awaiting for it to respond, which will ultimately change the scene.
Until that occurs, the player otherwise can unintentionally keep spamming the game.

There's no real reason for the timer to persist upon returning, especially at 5s,
and especially if it's (now) only ticking while this scene is active.