CN3UIBase::GetChildByID<T>(): Allow for inheritance
Realistically we should only ever explicitly fetch as the highest-most type,
but just in case, we should still allow for inheritance.

This is still faster than dynamic_cast(), and doesn't rely on RTTI.

Additionally, since we're using our very specific compile-time implementation now,
the compiler no longer needs to pass the UI type in the runtime call (though it could
still aggressively optimise it similarly before, but only really with PGO).

As a result, we get more precise, optimised output.