Default object visibility behaviour is flipped
We intend to *show* objects on an opt-in basis, not hide them.
This is why there's a desync with the official client regarding things like the anvil.
The anvil isn't being sent in the NPC info packet, so it's not shown in the official client.
It's only shown here because the default visibility state is flipped.

Additionally, officially we generate detailed meshes here, otherwise walking up to things like warp gates is a mess.

Refs #336