Fix inconsistent collision mesh generation
The core issue here is that it doesn't consistent rebuild the mesh after repositioning,
and at least in one case where it does, the zone hasn't changed yet so y isn't clamped
correctly for this zone.

Additionally, there's a couple of joint updates that aren't correctly triggered.
This is an official bug with 1.298 but we can correct it here as with newer client versions.

Resolves #55