Move into Server/network project
We decouple this from shared because it's used by the client.
The client has no business knowing about any of this, and will likely down the track
end up having its own asio implementation.

This would be largely incompatible as it's designed for a server.

It makes sense, then, to decouple this and have it in its own server-specific lib project.

Also: the shared output directory was using the project default. We should just force it
to $(RootDir)bin -- i.e. $(BaseOutDir) -- and just ensure we still reference that, even from
server projects.
I'm not sure how it got missed when the other output paths got updated.

Server\network resides in the server's bin dir (Server\bin) as it's server-specific.