Replace shared memory queue with boost::interprocess::message_queue
Aujard:
This previously tried to detect if Ebenezer dropped via the process ID in the queue,
and triggered saves if it wasn't found (which really didn't work at all).
This has been replaced with a heartbeat; Ebenezer will periodically send a heartbeat.
Instead of checking to see if the process exists, we now check the time since the last
heartbeat was received.

Ebenezer:
Use a mutex to guard access to region buffers; this previously shared SMQ defines and the
SMQ's "locking" implementation which really wasn't thread-safe at all.
This can be simplified to just a basic mutex, and we can throw away the SMQ defines.