Add SharedMemoryBlock to replace _USER_DATA MMF
This uses the raw internal type shared_memory_object.
This is a more direct translation than I was initially considering,
but it's the simplest and most straightforward for now.

Ideally we'd use managed_shared_memory and allocate our blocks there,
using it to setup more complex types, so we could use containers like maps.

For now this suffices as a fairly direct translation from raw MMF to
cross-platform Boost.Interprocess support.