MSBuild: Add support for VS2026 (v145)
When running VS2026 (there seems to be no good way to detect if a specific toolset actually exists at this point),
we now set the default platform toolset to v145 (2026).

Otherwise, we set it to v143 (2022) as before.

Additionally, replace the N3Base wrapper projects with their own individual implementations for client & tool builds.
It's a pain to debug with the wrapper project involved, so at the cost of having to duplicate (for the most part) our
config, it becomes a little more straightforward.

You can just debug the project directly at all times, even from All.sln.