Server: Implement headless mode (--headless)
Add p-ranav/argparse (v3.2) for command line parsing.

Attempt to detect terminals which definitely do not support FTXUI.
Unfortunately this doesn't include CLion, but I can't otherwise manage to detect it (there just doesn't seem to be a great way to detect it).
For this case, we can just enable the emulation setting in CLion or explicitly run it as headless.

 Resolves #664