Add basic FileReader & FileWriter leveraging llfio
Regular consumers will simply use the base File class and not need to be exposed to any of it.
Anything opening files (which is usually limited) will end up pulling in llfio through the reader or writer header.

I think it's better for the compiler to suffer in this case than the application via unnecessary heap allocations.