JpegFile: Use C file I/O API
We'd use <shared/FileReader.h> / <shared/FileWriter.h> but it currently requires a dependency to shared.lib and all that entails.
Since not all consumers of this use that, we'll just use the C file I/O API for this; on MSVC, this still performs considerably faster than the WinAPI.

In the future we might separate the file i/o to its own lib, then this won't be necessary, but it isn't really a big deal for this project.