CN3TableBase: Move as much out of the header as we can
Add CN3TableBaseImpl as a parent class to implement most of the shared logic that doesn't need to be templated.

MakeOffsetTable() could realistically be moved up a level, but this way the compiler can optimise it for the exact struct alignment.
This is typically only going to be one of 2 things, depending on whether or not it has 64-bit fields in them (in 32-bit builds, anyway),
but still.