Restore slightly more optimised path concatenation
operator=(const std::string& a, const std::string& b) will always allocate a new string.
Whereas with appending, it will only allocate a new string if there's not enough space reserved.

This will be largely useless later when this is all replaced, but for now let's not technically make performance worse
(even if the impact is negligible).