CMake: Add N3CE, force DX9 include paths via manually included prop sheets
DX9 is a mess. CMake doesn't play nicely with this.
It needs to be included after the regular include paths, but before everything else.
We can't really reliably guarantee that otherwise.

We cannot just use VS_GLOBAL_IncludePath as it's before VS defaults, which breaks their defaults.
The only other way to override IncludePath is from a property sheet via VS_USER_PROPS, which...
also just replaces the MS default. So we can only use it for this one specific thing (fine),
and never other property sheets.

Then of course this cannot be propagated, so we add require_directx9() to manually opt-in
to DirectX 9's property sheet (the other stuff is inherited just fine).